requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.1.post1
torch==2.1.2
transformers==4.37.2
shap==0.44.1
lime==0.2.0.1
openai==1.12.0
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.1
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slack-sdk==3.26.1
email-validator==2.1.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2
python-evtx==0.8.1
pywin32==306; sys_platform == 'win32'
winlogbeat==0.1.0; sys_platform == 'win32'

